Rename explainer module to explainers.
