+++ examples/readme.py
@@ -2,7 +2,7 @@ import numpy as np
 from sklearn.ensemble import RandomForestClassifier
 
 from ocean.datasets import load_adult
-from ocean.explainer import MixedIntegerProgramExplainer
+from ocean.explainers import MixedIntegerProgramExplainer
 
 # Load the adult dataset
 (data, target), mapper = load_adult()
